+++ stable_baselines3/common/buffers.py
@@ -279,7 +279,7 @@ class ReplayBuffer(BaseBuffer):
         :param batch_size: Number of element to sample
         :param env: associated gym VecEnv
             to normalize the observations/rewards when sampling
-        :return:
+        :return: a batch of sampled experiences from the buffer.
         """
         if not self.optimize_memory_usage:
             return super().sample(batch_size=batch_size, env=env)